+++ migrations/20260117171127_create_users.sql
@@ -1,0 +1,9 @@
+-- up
+CREATE TABLE users (
+    id INTEGER PRIMARY KEY,
+    username TEXT NOT NULL,
+    email TEXT NOT NULL
+);
+
+-- down
+DROP TABLE users;